Add trade bootstrap stress test and fix dashboard readability
- Add run_bootstrap() to strategies/robustness.py: reshuffles realized
  trade P&Ls to test if drawdown was from unlucky ordering or inherent risk
- Add bootstrap_drawdown_chart() histogram to app/charts.py
- Integrate bootstrap section in Strategy Dashboard between MC and regime
- Fix metric cards: compact dollar formatting ($1.23M), no arrow on Max DD
- Fix x-axis clutter on all time-series charts: date type with nticks/tickformat
- Fix price_with_signals: reduce 59 empty annotations to 2 (Long/Short)
- Cap backtests to 2025 only (exclude 2026 data)
- Add *.pdf to .gitignore
- Exclude robustness.py from strategy discovery
- Integer position sizing in backtest engine
- 4 new tests (3 bootstrap + 1 chart)

+++ strategies/backtest.py
@@ -76,7 +76,7 @@ def compute_daily_pnl(
             # Size new position from current equity
             if pos != 0:
                 allocation = current_equity * risk_pct
-                current_units = allocation / closes[i] if closes[i] > 0 else 0.0
+                current_units = int(allocation / closes[i]) if closes[i] > 0 else 0
             else:
                 current_units = 0.0
